Refactor: 자바 표준 API와 입력 정규화 정리

+++ application/src/main/java/com/personal/happygallery/application/batch/BatchLoggingAspect.java
@@ -43,7 +43,7 @@ public class BatchLoggingAspect {
                                 durationMs);
                     }
                 } else if (result instanceof Number count) {
-                    log.info("[배치] {} 완료: {}건 ({}ms)", jobName, count.intValue(), durationMs);
+                    log.info("[배치] {} 완료: {}건 ({}ms)", jobName, count, durationMs);
                 } else {
                     log.info("[배치] {} 완료 ({}ms)", jobName, durationMs);
                 }
